feat: pulumi provider
The pulumi provider wraps the pulumi CLI for deploying and destroying
pulumi stacks.

The pulumi module type supports merging variables and varfiles with
pulumi configs (via the `pulumiVariables` and `pulumiVarfiles` config
fields). Pulumi modules that deploy sensitive resources can be protected
by setting `autoApply: false` and/or `allowDestroy: false`.

When the `deployFromPreview` config variable is provided, the stack will
be deployed with pre-generated plans and configs, which is useful for
two-phase CI flows (where the proposed plans can be reviewed and then
passed to the deployment pipeline).

Also included are plugin commands for cancelling, refreshing, and
reimporting stacks (in dependency order, when called for more than one
service).

Finally, we've also moved the terraform plugin out of the Core package
into the `plugins` directory.

TODO: Write integration tests and a usage guide.